removed Monitor class
git-svn-id: http://konoha.googlecode.com/svn/trunk/konoha@294 11e8e06f-943d-0410-b0b4-4be84c385469

+++ src/lang/vmcodelibs.c
@@ -609,7 +609,7 @@ void knh_opcode_shift(knh_opline_t *c, int shift)
 #define JUMP        L_HEAD
 #define TC(c)
 #define DISPATCH_START(pc) L_HEAD:;switch(pc->opcode) {
-#define DISPATCH_END(pc)   } KNH_DIE("unknown opcode=%d", pc->opcode); 
+#define DISPATCH_END(pc)   } KNH_DIE("unknown opcode=%d", (int)pc->opcode); 
 #define GOTO_PC(pc)         GOTO_NEXT()
 #endif/*K_USING_THCODE_*/
